[tegra] add MJPEG stream support
This is proof-of-concept code that configures the camera as before,
but also allows MJPEG streams be built.

In the example, raw camera and CV output is streamed. This can be
used for input to the drive station of either raw camera or showing
CV frames (e.g. like the green contours).

This could also replace the imshow() calls in cv_capture and just
use the local web browser (e.g. http:://127.0.0.1:1185) to view
the output (one less thing to change when it runs on the robot).