updated deadband to use original joystick values instead of two different values, new shooter values for speed
we were using two different values from the joysticks originally - one to cube/perform calculations on, and one for the deadband. we now have an original left Y value that we use for the deadband and then a new variable with the calculations on it so that we don't use two different values where we should be using just 1. we also decreased the value of the deadband and slowed down shooter values for the rollers.